+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Chat Patches `214.6.13` for Minecraft 1.21.2-1.21.4 on Fabric, Quilt
+- Added the dynamic shift feature, which shifts the chat according to the player's health and armor bars (thanks [radioactive-exe](https://github.com/radioactive-exe)!)
+
 ## Chat Patches `214.6.12` for Minecraft 1.21.2-1.21.4 on Fabric, Quilt
 - Bumped dependencies, now requires Fabric Loader >=0.16.10
 - Fixed some playernames not formatting properly